Move interceptor specific classes to interceptor package
Move interceptor specific classes to interceptor package

#3485
