+++ axon-5/api-changes.md
@@ -2016,7 +2016,7 @@ This section contains five tables:
 | org.axonframework.messaging.core.annotation.AnnotatedHandlerAttributes                                 | org.axonframework.messaging.core.annotation.AnnotatedHandlerAttributes                                         | No                             |
 | org.axonframework.messaging.core.annotation.AnnotatedHandlerInspector                                  | org.axonframework.messaging.core.annotation.AnnotatedHandlerInspector                                          | No                             |
 | org.axonframework.messaging.core.annotation.AnnotatedMessageHandlingMemberDefinition                   | org.axonframework.messaging.core.annotation.AnnotatedMessageHandlingMemberDefinition                           | No                             |
-| org.axonframework.messaging.core.annotation.ChainedMessageHandlerInterceptorMember                     | org.axonframework.messaging.core.annotation.ChainedMessageHandlerInterceptorMember                             | No                             |
+| org.axonframework.messaging.core.annotation.ChainedMessageHandlerInterceptorMember                     | org.axonframework.messaging.core.interception.annotation.ChainedMessageHandlerInterceptorMember                | No                             |
 | org.axonframework.messaging.core.annotation.ClasspathHandlerDefinition                                 | org.axonframework.messaging.core.annotation.ClasspathHandlerDefinition                                         | No                             |
 | org.axonframework.messaging.core.annotation.ClasspathHandlerEnhancerDefinition                         | org.axonframework.messaging.core.annotation.ClasspathHandlerEnhancerDefinition                                 | No                             |
 | org.axonframework.messaging.core.annotation.ClasspathParameterResolverFactory                          | org.axonframework.messaging.core.annotation.ClasspathParameterResolverFactory                                  | No                             |
@@ -2027,7 +2027,7 @@ This section contains five tables:
 | org.axonframework.messaging.core.annotation.HandlerDefinition                                          | org.axonframework.messaging.core.annotation.HandlerDefinition                                                  | No                             |
 | org.axonframework.messaging.core.annotation.HandlerEnhancerDefinition                                  | org.axonframework.messaging.core.annotation.HandlerEnhancerDefinition                                          | No                             |
 | org.axonframework.messaging.core.annotation.HasHandlerAttributes                                       | org.axonframework.messaging.core.annotation.HasHandlerAttributes                                               | No                             |
-| org.axonframework.messaging.core.annotation.InterceptorChainParameterResolverFactory                   | org.axonframework.messaging.core.annotation.InterceptorChainParameterResolverFactory                           | No                             |
+| org.axonframework.messaging.core.annotation.InterceptorChainParameterResolverFactory                   | org.axonframework.messaging.core.interception.annotation.InterceptorChainParameterResolverFactory              | No                             |
 | org.axonframework.messaging.core.annotation.MessageHandler                                             | org.axonframework.messaging.core.annotation.MessageHandler                                                     | No                             |
 | org.axonframework.messaging.core.annotation.MessageHandlerInterceptorDefinition                        | org.axonframework.messaging.core.annotation.interceptors.MessageHandlerInterceptorDefinition                   | No                             |
 | org.axonframework.messaging.core.annotation.MessageHandlerInterceptorMemberChain                       | org.axonframework.messaging.core.annotation.interceptors.MessageHandlerInterceptorMemberChain                  | No                             |
