+++ messaging/src/main/resources/META-INF/services/org.axonframework.messaging.core.annotation.ParameterResolverFactory
@@ -21,8 +21,7 @@ org.axonframework.messaging.eventhandling.replay.annotation.ReplayContextParamet
 org.axonframework.messaging.eventhandling.replay.annotation.ReplayParameterResolverFactory
 org.axonframework.messaging.core.annotation.AggregateTypeParameterResolverFactory
 org.axonframework.messaging.core.annotation.DefaultParameterResolverFactory
-org.axonframework.messaging.deadletter.DeadLetterParameterResolverFactory
-org.axonframework.messaging.core.annotation.InterceptorChainParameterResolverFactory
+org.axonframework.messaging.core.interception.annotation.InterceptorChainParameterResolverFactory
 org.axonframework.messaging.core.annotation.MessageIdentifierParameterResolverFactory
 org.axonframework.messaging.core.interception.annotation.ResultParameterResolverFactory
 org.axonframework.messaging.core.annotation.SourceIdParameterResolverFactory
