+++ stash/legacy-aggregate/src/main/java/org/axonframework/modelling/command/inspection/AnnotatedCommandHandlerInterceptor.java
@@ -20,7 +20,7 @@ import org.axonframework.messaging.commandhandling.CommandMessage;
 import org.axonframework.messaging.core.MessageHandlerInterceptorChain;
 import org.axonframework.messaging.core.MessageHandlerInterceptor;
 import org.axonframework.messaging.core.MessageStream;
-import org.axonframework.messaging.core.annotation.InterceptorChainParameterResolverFactory;
+import org.axonframework.messaging.core.interception.annotation.InterceptorChainParameterResolverFactory;
 import org.axonframework.messaging.core.annotation.MessageHandlingMember;
 import org.axonframework.messaging.core.unitofwork.ProcessingContext;
